Refactoring Community Page Component Structure and CSS
