+++ client/src/components/app/AppHeader.js
@@ -1,4 +1,4 @@
-import { Link, Route, Switch, useRouteMatch } from "react-router-dom";
+import { Link, Route, Switch } from "react-router-dom";
 
 // import BrandLogo from "../../assets/2973980108_ed69085414_o.jpg";
 import { BsGrid3X3Gap } from "react-icons/bs";
@@ -6,8 +6,8 @@ import { FaPlus } from "react-icons/fa";
 import { FaUserAlt } from "react-icons/fa";
 
 import Modal from "../modal/Modal.js";
-import NewCommunityPage from "../../pages/community/NewCommunityPage.js";
-import EditCommunityPage from "../../pages/community/EditCommunityPage.js";
+import NewCommunityPage from "../../pages/communities/NewCommunityPage.js";
+import EditCommunityPage from "../../pages/communities/EditCommunityPage.js";
 
 const ADD_ACTION_FORM_ID = "newCommunity";
 const UPDATE_ACTION_FORM_ID = "editCommunity";
