--- client/src/domain/Community/Profile.css
@@ -1,98 +1,0 @@
-section.CommunityProfile {
-  flex: 1 1 auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  width: 100%;
-  position: relative;
-  overflow: hidden;
-  background-color: var(--card-bg-color);
-}
-
-section.CommunityProfile .CommunityProfile__CommunityName {
-  display: none;
-  width: 100%;
-  border-bottom: 4px solid lightblue;
-
-  font-size: 1rem;
-  font-weight: normal;
-}
-
-section.CommunityProfile .CommunityProfile__Image--wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  padding: 0;
-  width: 50px;
-  height: 50px;
-}
-section.CommunityProfile .CommunityProfile__Image--wrapper img {
-  display: block;
-  object-fit: cover;
-  border-radius: 10px;
-  width: 100%;
-  height: 100%;
-}
-
-section.CommunityProfile .CommunityProfile__CommunityMeta {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  color: darkslategray;
-  width: 100%;
-  letter-spacing: 1px;
-  padding: 1rem;
-
-  font-size: 0.75rem;
-  display: none;
-}
-
-section.CommunityProfile .CommunityProfile__CommunityMeta li {
-  display: flex;
-  align-items: center;
-}
-section.CommunityProfile .CommunityProfile__CommunityMeta .CommunityMeta__Icon {
-  margin-right: 0.5rem;
-}
-
-section.CommunityProfile
-  .CommunityProfile__CommunityMeta
-  .CommunityMeta__Caption {
-  display: none;
-}
-
-section.CommunityProfile .CommunityProfile__CommunityMeta .OnlineIcon {
-  margin-right: 0.25rem;
-  margin-left: 0.25rem;
-  font-size: 80%;
-  color: green;
-}
-
-/* MEDIA QUERIES */
-
-@media all and (min-width: 600px) {
-  section.CommunityProfile .CommunityProfile__CommunityName {
-    display: block;
-  }
-
-  section.CommunityProfile .CommunityProfile__Image--wrapper {
-    width: 100%;
-    height: auto;
-  }
-}
-@media all and (min-width: 800px) {
-  section.CommunityProfile .CommunityProfile__CommunityMeta {
-    display: flex;
-    font-size: 1rem;
-  }
-}
-
-@media all and (min-width: 1000px) {
-  section.CommunityProfile
-    .CommunityProfile__CommunityMeta
-    .CommunityMeta__Caption {
-    display: inline-block;
-    margin-right: 0.25rem;
-  }
-}
