+++ client/src/layouts/CommunityLayout.css
@@ -12,215 +12,9 @@ main.CommunityLayout {
   color: var(--main-font-color);
   background-color: var(--main-bg-color);
 }
-/* default: hidden */
-main.CommunityLayout .CommunitySidebar {
-  background-color: bisque;
-  flex-basis: 33%;
-  flex-grow: 1;
-  visibility: hidden;
-  flex-wrap: nowrap;
+
+:root {
+  --CommunitySidebar-bg-color: #b5cfd8;
+  --CommunityMain__Header-bg-color: #e8ecf1;
+  --CommunityMainHeader__Item-bg-color: var(--CommunitySidebar-bg-color);
 }
-
-main.CommunityLayout .CommunityMain {
-  background-color: rgb(119, 115, 111);
-  flex: 1 1 auto;
-
-  min-width: 100%;
-
-  /* flex-basis: 0;
-  flex-grow: 999; */
-}
-
-/* main.CommunityLayout .CommunitySidebar--hidden {
-  visibility: hidden;
-} */
-main.CommunityLayout .CommunitySidebar--toggled {
-  visibility: visible;
-}
-
-main.CommunityLayout .CommunitySidebar--toggled + .CommunityMain {
-  min-width: 50%;
-}
-
-main.CommunityLayout .CommunitySidebar .CommunitySidebar__CommunityHeader {
-  flex: 1 1 auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  width: 100%;
-  position: relative;
-  overflow: hidden;
-  box-shadow: rgb(50 50 93 / 25%) 0px 13px 27px -5px,
-    rgb(0 0 0 / 30%) 0px 8px 16px -8px;
-  background-color: var(--card-bg-color);
-  border: 2px solid whitesmoke;
-  transition: all 0.3s ease-in-out;
-  -webkit-transform-style: preserve-3d;
-  -webkit-backface-visibility: hidden;
-  /* visibility: hidden; */
-  display: none;
-}
-
-main section.Community .Community__ProfilePic-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  padding: 0;
-  width: 100%;
-}
-
-main section.Community .Community__ProfilePic-wrapper img {
-  display: block;
-
-  object-fit: cover;
-  border-radius: 10px;
-}
-
-main.CommunityLayout .CommunityMain {
-  display: flex;
-  flex-direction: column;
-}
-main.CommunityLayout .CommunityMain .CommunityMain__Header {
-  display: flex;
-  /* flex-grow: 1; */
-  flex-basis: 50px;
-  background-color: chocolate;
-}
-main.CommunityLayout .CommunityMain .CommunityMain__Content {
-  display: flex;
-  flex: 1 1 auto;
-  min-width: 100%;
-}
-
-/* CSS PROPERTIES TO CONTROL RESPONSIVENESS  */
-
-/*
-  --xs: 400px;
-  --sm: 576px;
-  --md: 768px;
-  --lg: 992px;
-  --xl: 1024px;
-  --xxl: 1200px;
-  --hd: 1400px;   */
-
-@media all and (min-width: 400px) {
-  /* default: collapsed to min bar  */
-  main.CommunityLayout .CommunitySidebar {
-    visibility: visible;
-    max-width: 55px;
-  }
-
-  main.CommunityLayout .CommunityMain {
-    min-width: 50%;
-  }
-
-  /* toggled: extend to max */
-  main.CommunityLayout .CommunitySidebar--toggled {
-    visibility: visible;
-    max-width: 33%;
-  }
-
-  main.CommunityLayout .CommunitySidebar--toggled + .CommunityMain {
-    min-width: 50%;
-  }
-}
-
-@media all and (min-width: 576px) {
-  /* default: collapsed to min bar  */
-  main.CommunityLayout .CommunitySidebar {
-    max-width: 75px;
-  }
-  /* toggled: extend to max */
-  main.CommunityLayout .CommunitySidebar--toggled {
-    max-width: 33%;
-  }
-}
-
-@media all and (min-width: 768px) {
-  /* default: collapsed to min bar  */
-  main.CommunityLayout .CommunitySidebar {
-    max-width: 75px;
-  }
-  /* toggled: extend to max */
-  main.CommunityLayout .CommunitySidebar--toggled {
-    max-width: 33%;
-  }
-}
-
-@media all and (min-width: 992px) {
-  /* default: collapsed to min bar  */
-  main.CommunityLayout .CommunitySidebar {
-    width: 33%;
-    max-width: 275px;
-  }
-  /* toggled: extend to max */
-  main.CommunityLayout .CommunitySidebar--toggled {
-    max-width: 75px;
-  }
-}
-
-@media all and (min-width: 1024px) {
-  /* default: collapsed to min bar  */
-  main.CommunityLayout .CommunitySidebar {
-    width: 33%;
-    max-width: 300px;
-  }
-  /* toggled: extend to max */
-  main.CommunityLayout .CommunitySidebar--toggled {
-    max-width: 75px;
-  }
-}
-@media all and (min-width: 1200px) {
-  /* default: collapsed to min bar  */
-  main.CommunityLayout .CommunitySidebar {
-    width: 33%;
-    max-width: 350px;
-  }
-  /* toggled: extend to max */
-  main.CommunityLayout .CommunitySidebar--toggled {
-    max-width: 75px;
-  }
-}
-
-@media all and (min-width: 1400px) {
-  /* default: collapsed to min bar  */
-  main.CommunityLayout .CommunitySidebar {
-    width: 33%;
-    max-width: 400px;
-  }
-  /* toggled: extend to max */
-  main.CommunityLayout .CommunitySidebar--toggled {
-    max-width: 75px;
-  }
-}
-
-/*
-@media all and (min-width: 600px) {
-  main.CommunityLayout {
-    flex-wrap: wrap;
-  }
-
-  main.CommunityLayout .CommunitySidebar {
-    background-color: bisque;
-
-    flex-basis: 25%;
-    max-width: 33%;
-    flex-grow: 1;
-    visibility: visible;
-  }
-  main.CommunityLayout .CommunitySidebar--toggled {
-  }
-  main.CommunityLayout .CommunitySidebar--fadedout {
-  }
-
-  main.CommunityLayout .CommunitySidebar .CommunitySidebar__CommunityHeader {
-    display: flex;
-  }
-
-  main.CommunityLayout .CommunityMain {
-    background-color: rgb(119, 115, 111);
-    flex: 1 1 auto;
-    min-width: 50%;
-  }
-} */
