+++ client/src/pages/communities/Community/Main.css
@@ -1,0 +1,33 @@
+main.CommunityLayout .CommunityMain .CommunityMain__Header {
+  display: flex;
+  flex-basis: 50px;
+
+  background-color: var(--CommunityMain__Header-bg-color);
+
+  gap: 1rem;
+  padding: 0.25rem 0.25rem 0.25rem 0rem;
+
+  align-items: center;
+  align-content: center;
+  justify-content: flex-start;
+}
+main.CommunityLayout .CommunityMain .CommunityMain__Content {
+  display: flex;
+  flex: 1 1 auto;
+  min-width: 100%;
+}
+
+main.CommunityLayout .CommunityMain a.CommunityMainHeader__Item {
+  padding: 0.3rem;
+  padding-right: 0.75rem;
+
+  background-color: #b5cfd8;
+  border-top-right-radius: 10px;
+  border-bottom-right-radius: 10px;
+}
+main.CommunityLayout
+  .CommunityMain
+  a.CommunityMainHeader__Item
+  .sidebarToggleIcon {
+  color: whitesmoke;
+}
