+++ client/src/pages/communities/Community/Sidebar/Profile.css
@@ -1,0 +1,96 @@
+main.CommunityLayout .CommunitySidebar .CommunityProfile {
+  flex: 1 1 auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  width: 100%;
+  position: relative;
+  overflow: hidden;
+  background-color: var(--card-bg-color);
+}
+
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__CommunityName {
+  display: none;
+  width: 100%;
+  border-bottom: 4px solid lightblue;
+  font-size: 1rem;
+  font-weight: normal;
+}
+
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__Image--wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  padding: 0;
+  max-width: auto;
+  max-height: auto;
+}
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__Image--wrapper
+  img {
+  display: block;
+  object-fit: cover;
+  border-radius: 10px;
+  width: 100%;
+  height: 100%;
+}
+
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__CommunityMeta {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  color: darkslategray;
+  width: 100%;
+  letter-spacing: 1px;
+  padding: 1rem;
+
+  font-size: 0.75rem;
+  display: none;
+}
+
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__CommunityMeta
+  li {
+  display: flex;
+  align-items: center;
+}
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__CommunityMeta
+  .CommunityMeta__Icon {
+  margin-right: 0.5rem;
+}
+
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__CommunityMeta
+  .CommunityMeta__Caption {
+  display: none;
+}
+
+main.CommunityLayout
+  .CommunitySidebar
+  .CommunityProfile
+  .CommunityProfile__CommunityMeta
+  .OnlineIcon {
+  margin-right: 0.25rem;
+  margin-left: 0.25rem;
+  font-size: 80%;
+  color: green;
+}
