+++ client/src/pages/communities/Community/Sidebar/Profile.js
@@ -1,22 +1,8 @@
-import "./Community.css";
-import CommunityFallbackProfilePic from "../../assets/happy-students.jpg";
+import CommunityFallbackProfilePic from "../../../../assets/happy-students.jpg";
 import "./Profile.css";
 import { FaUserAlt } from "react-icons/fa";
 import { FaDotCircle } from "react-icons/fa";
 import { FaUsers } from "react-icons/fa";
-// import CommunitiesContext from "../../contexts/CommunitiesContext";
-
-// import { useContext } from "react";
-import { useHistory } from "react-router";
-
-// import { Link } from "react-router-dom";
-// import { FaRegEdit, FaTrashAlt } from "react-icons/fa";
-// import axios from "axios";
-
-// TODO: Get rid of that conditional wrapper again - should be possible to
-// just set an onLick-handler on the section-tag itself and basta kanzla!
-// const ConditionalWrapper = ({ condition, wrapper, children }) =>
-//   condition ? wrapper(children) : children;
 
 export default function Profile({ community }) {
   const communityProfilePicImgSrc =
@@ -28,9 +14,6 @@ export default function Profile({ community }) {
       key={community.id}
       id={community.id}
     >
-      <h3 className="CommunityProfile__CommunityName truncate">
-        {community.name}
-      </h3>
       <p className="CommunityProfile__Image--wrapper">
         <img
           src={communityProfilePicImgSrc}
@@ -39,7 +22,7 @@ export default function Profile({ community }) {
         />
       </p>
 
-      <h3 className="CommunityProfile__CommunityName truncate truncate">
+      <h3 className="CommunityProfile__CommunityName truncate ">
         {community.name}
       </h3>
